refactor: near-contract-transport crate for NEAR call/view vocabulary
Renames mpc-call-args and moves ViewArgs, ViewContract, BlockHeight and
ObservedState there. ObservedState becomes generic over the height witness
(ObservedState<T, H = BlockHeight>) and ViewContract gains type ObservedAt,
so backends that cannot report a height declare ObservedAt = () instead of
faking one. Trait-side dependencies sit behind an opt-in traits feature;
the wasm test contract stays payload-only. chain-gateway re-exports none of
the moved types: mpc-node, tee-context and the integration tests depend on
near-contract-transport directly.
