+++ crates/chain-gateway-test-contract/Cargo.toml
@@ -19,7 +19,7 @@ crate-type = ["cdylib", "lib"]
 
 [dependencies]
 borsh = { workspace = true }
-mpc-call-args = { workspace = true }
+near-contract-transport = { workspace = true }
 near-sdk = { workspace = true }
 serde_json = { workspace = true }
