+++ crates/chain-gateway/src/event_subscriber/block_events.rs
@@ -2,10 +2,8 @@ use derive_more::{Deref, From};
 use near_account_id::AccountId;
 use near_indexer_primitives::CryptoHash;
 
-use crate::{
-    event_subscriber::recent_blocks_tracker::BlockStatusHandle,
-    types::{BlockEntropy, BlockHeight},
-};
+use crate::{event_subscriber::recent_blocks_tracker::BlockStatusHandle, types::BlockEntropy};
+use near_contract_transport::BlockHeight;
 
 /// The BlockUpdate returned by the Chain indexer.
 /// Similar to [`ChainBlockUpdate`](../../node/src/indexer/handler.rs) in the `mpc-node` crate.
