+++ crates/chain-gateway/src/types.rs
@@ -1,54 +1,11 @@
-use derive_more::{Display, From, Into};
+use derive_more::Into;
 use near_indexer_primitives::CryptoHash;
 use near_indexer_primitives::types::Gas;
-use serde::{Deserialize, Serialize, de::DeserializeOwned};
 
-pub use mpc_call_args::FunctionCallArgs;
-use mpc_call_args::NearGas;
-
-use crate::errors::ChainGatewayError;
+use near_contract_transport::{NearGas, ObservedState};
 
 pub(crate) fn to_action_gas(gas: NearGas) -> Gas {
     Gas::from_gas(gas.as_gas())
-}
-
-/// A NEAR view-function request: method name and encoded args.
-#[derive(Debug, Clone)]
-pub struct ViewArgs {
-    pub method_name: String,
-    pub args: Vec<u8>,
-}
-
-impl ViewArgs {
-    pub fn new(method_name: impl Into<String>, args: Vec<u8>) -> Self {
-        Self {
-            method_name: method_name.into(),
-            args,
-        }
-    }
-
-    /// A view call taking no arguments (an empty JSON object).
-    pub fn no_args(method_name: impl Into<String>) -> Self {
-        Self::new(method_name, b"{}".to_vec())
-    }
-}
-
-#[derive(
-    Into, From, Copy, Clone, Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord, Debug, Display,
-)]
-pub struct BlockHeight(u64);
-
-impl BlockHeight {
-    pub fn saturating_add(self, delta: u64) -> Self {
-        BlockHeight(self.0.saturating_add(delta))
-    }
-    pub fn saturating_sub(self, delta: u64) -> Self {
-        BlockHeight(self.0.saturating_sub(delta))
-    }
-    /// Block distance from `earlier` to `self`. Saturates to `0` if `earlier > self`.
-    pub fn blocks_since(self, earlier: BlockHeight) -> u64 {
-        self.0.saturating_sub(earlier.0)
-    }
 }
 
 #[derive(Clone, Into, Debug)]
@@ -66,73 +23,5 @@ impl From<CryptoHash> for BlockEntropy {
     }
 }
 
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct ObservedState<T = Vec<u8>> {
-    pub observed_at: BlockHeight,
-    pub value: T,
-}
-
 /// block height and block hash
 pub type LatestFinalBlockInfo = ObservedState<CryptoHash>;
-
-impl ObservedState<Vec<u8>> {
-    pub fn deserialize<Res: DeserializeOwned>(
-        self,
-    ) -> Result<ObservedState<Res>, ChainGatewayError> {
-        let value = serde_json::from_slice::<Res>(&self.value).map_err(|err| {
-            ChainGatewayError::Deserialization {
-                message: err.to_string(),
-            }
-        })?;
-        Ok(ObservedState {
-            observed_at: self.observed_at,
-            value,
-        })
-    }
-}
-
-#[cfg(test)]
-#[expect(non_snake_case)]
-mod tests {
-    use assert_matches::assert_matches;
-    use serde::Deserialize;
-
-    use crate::{
-        errors::ChainGatewayError,
-        types::{ObservedState, ViewArgs},
-    };
-
-    #[test]
-    fn no_args__should_encode_an_empty_json_object() {
-        assert_eq!(ViewArgs::no_args("m").args, b"{}");
-    }
-
-    #[derive(Debug, Deserialize, PartialEq, Eq)]
-    struct Num {
-        a: u32,
-    }
-
-    #[test]
-    fn test_deserialize_ok() {
-        let observed = ObservedState {
-            observed_at: 7.into(),
-            value: br#"{"a":1}"#.to_vec(),
-        };
-
-        let typed: ObservedState<Num> = observed.deserialize().unwrap();
-
-        assert_eq!(typed.observed_at, 7.into());
-        assert_eq!(typed.value, Num { a: 1 });
-    }
-
-    #[test]
-    fn test_deserialize_err() {
-        let observed = ObservedState {
-            observed_at: 7.into(),
-            value: b"not json".to_vec(),
-        };
-
-        let err = observed.deserialize::<Num>().unwrap_err();
-        assert_matches!(err, ChainGatewayError::Deserialization { .. });
-    }
-}
