--- crates/mpc-call-args/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "mpc-call-args"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-description = "Receiver-less NEAR function-call descriptor (method, args, gas, deposit) shared across the MPC stack"
-
-[dependencies]
-near-gas = { workspace = true }
-near-token = { workspace = true }
-
-[lints]
-workspace = true
