+++ crates/near-contract-transport/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "near-contract-transport"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "NEAR function-call payload and the CallContract/ViewContract transport traits shared across the MPC stack"
+
+[features]
+traits = ["dep:near-account-id"]
+
+[dependencies]
+derive_more = { workspace = true }
+near-account-id = { workspace = true, optional = true }
+near-gas = { workspace = true }
+near-token = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+
+[lints]
+workspace = true
