+++ crates/node/Cargo.toml
@@ -46,6 +46,7 @@ near-account-id = { workspace = true }
 near-async = { workspace = true }
 near-client = { workspace = true }
 near-config-utils = { workspace = true }
+near-contract-transport = { workspace = true }
 near-crypto = { workspace = true }
 near-indexer = { workspace = true }
 near-indexer-primitives = { workspace = true }
