+++ crates/node/src/requests/queue.rs
@@ -4,8 +4,8 @@ use crate::primitives::ParticipantId;
 use crate::requests::metrics;
 use crate::types::{self, Request, RequestId, RequestsUpdate};
 use chain_gateway::event_subscriber::recent_blocks_tracker::BlockStatusHandle;
-use chain_gateway::types::BlockHeight;
 use k256::sha2::Sha256;
+use near_contract_transport::BlockHeight;
 use near_indexer_primitives::CryptoHash;
 use near_indexer_primitives::types::NumBlocks;
 use near_time::Duration;
