+++ crates/tee-context/Cargo.toml
@@ -7,9 +7,9 @@ license = { workspace = true }
 [dependencies]
 chain-gateway = { workspace = true }
 mpc-primitives = { workspace = true }
-near-mpc-contract-interface = { workspace = true, features = ["call-args"] }
-
 near-account-id = { workspace = true }
+near-contract-transport = { workspace = true }
+near-mpc-contract-interface = { workspace = true, features = ["call-args"] }
 serde = { workspace = true }
 serde_json = { workspace = true }
 thiserror = { workspace = true }
